+++ apps/chronicles/navigation/index.tsx
@@ -35,6 +35,7 @@ import { PortalScreen } from "../screens/components/PortalScreen";
 import { DialogScreen } from "../screens/components/DialogScreen";
 import { EnvironmentScreen } from "../screens/components/EnvironmentScreen";
 import { ButtonCellScreen } from "../screens/components/ButtonCellScreen";
+import { SwitchCellScreen } from "../screens/components/SwitchCellScreen";
 import { trackNavigation } from "@equinor/mad-insights";
 import { SwitchScreen } from "../screens/components/SwitchScreen";
 import { LoginScreen } from "../screens/LoginScreen";
@@ -120,6 +121,7 @@ function DiscoverNavigator() {
             <DiscoverStack.Screen name="Cell" component={CellScreen} />
             <DiscoverStack.Screen name="NavigationCell" component={NavigationCellScreen} />
             <DiscoverStack.Screen name="ButtonCell" component={ButtonCellScreen} />
+            <DiscoverStack.Screen name="SwitchCell" component={SwitchCellScreen} />
             <DiscoverStack.Screen name="Accordion" component={AccordionScreen} />
             <DiscoverStack.Screen name="Menu" component={MenuScreen} />
             <DiscoverStack.Screen