+++ apps/chronicles/screens/components/CellScreen.tsx
@@ -39,7 +39,7 @@ export const CellScreen = () => {
                 </Typography>
             </View>
             <Spacer />
-            <Cell style={{ height: 60 }} />
+            <Cell />
             <Spacer />
             <View style={styles.readableContent}>
                 <Typography variant="h2">Groups</Typography>
@@ -51,26 +51,32 @@ export const CellScreen = () => {
             </View>
             <Spacer />
             <Cell.Group title="My first cell group">
-                <Cell style={{ height: 60 }} />
-                <Cell style={{ height: 60 }} />
-                <Cell style={{ height: 60 }} />
+                <Cell />
+                <Cell />
+                <Cell />
             </Cell.Group>
             <Spacer />
             <View style={styles.readableContent}>
                 <Typography>
                     The clumping behavior applies for any type of cell. The group below contains
-                    three cells: one navigation cell, one normal blank cell and one custom cell made
-                    for this screen. You would typically create app specific cells for your apps
-                    needs.
+                    three cells: one switch cell, one navigation cell and one custom cell made for
+                    this screen. You would typically create app specific cells for your apps needs.
                 </Typography>
             </View>
             <Spacer />
             <Cell.Group>
-                <Cell.Navigation title="This is a navigation cell" onPress={() => null} />
-                <Cell>
-                    <Typography>This is a normal blank cell</Typography>
-                </Cell>
-                <MyCustomCell title="This is a custom cell" />
+                <Cell.Switch
+                    title="This cell is a switch cell"
+                    isActive={false}
+                    onChange={() => null}
+                />
+                <Cell.Navigation
+                    title="This is a navigation cell"
+                    description="And this is its description!"
+                    iconName="navigation-outline"
+                    onPress={() => null}
+                />
+                <MyCustomCell title="And this is a custom cell" />
             </Cell.Group>
             <Spacer />
             <View style={styles.readableContent}>
@@ -85,12 +91,11 @@ export const CellScreen = () => {
                 <Cell
                     leftAdornment={
                         <View style={styles.adornment}>
-                            <Typography> This is the left adornment</Typography>
+                            <Typography>This is the left adornment</Typography>
                         </View>
                     }
-                    style={{ height: 60 }}
                 />
-                <Cell style={{ height: 60 }}>
+                <Cell>
                     <View style={styles.child}>
                         <Typography>This element is a child of Cell</Typography>
                     </View>
@@ -106,7 +111,6 @@ export const CellScreen = () => {
                             <Typography> Right</Typography>
                         </View>
                     }
-                    style={{ height: 60 }}
                 >
                     <View style={styles.child}>
                         <Typography>Main content</Typography>
@@ -122,10 +126,8 @@ export const CellScreen = () => {
                 </Typography>
             </View>
             <Spacer />
-            <Cell onPress={() => null} style={{ height: 60 }}>
-                <View style={{ justifyContent: "center", flex: 1 }}>
-                    <Typography>This cell responds to touch!</Typography>
-                </View>
+            <Cell onPress={() => null}>
+                <Typography>This cell responds to touch!</Typography>
             </Cell>
             <Spacer />
             <View style={styles.readableContent}>
@@ -212,7 +214,6 @@ const themeStyles = EDSStyleSheet.create(theme => ({
         borderRadius: theme.geometry.border.elementBorderRadius,
     },
     adornment: {
-        flex: 1,
         borderWidth: 2,
         justifyContent: "center",
         alignItems: "center",